Reorganize main file
